backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic-settings==2.5.2
sqlalchemy==2.0.35
psycopg[binary]==3.3.3
pgvector==0.3.5
alembic==1.13.2
pymupdf==1.27.2.2
python-docx==1.1.2
tiktoken==0.12.0
sentence-transformers==3.1.1
httpx==0.27.2
pytest==8.3.3
